Prevent additional faults from being injected
To my understanding, the 'injectFault' function runs repeatedly for every type of power fault, but only injects them if they are activated at the time. I added logic to only allow the first fault activated to go through and be injected; any power faults that are activated afterwards skip injection and thus should not affect the simulation. The user is warned when this occurs.

Additionally, if the user deactivates the original fault and then tries to reactivate it, it will still not work. Only one power fault activation per simulation is allowed.