Experimental prevention of m_faults from being updated
I'm unsure if these changes are necessary to the OCEANWATER-894 ticket, as I'm unsure as to what the full role of m_faults is. Essentially, these changes (like the previous commit) also detect if a fault has been activated, and if faults is copied over to m_faults with a power fault that should not be activated, it is (theoretically) deactivated before the end of the function. I'm not particularly confident in this commit; it may not be necessary at all based on the changes to FaultInjector.cpp.