Implemented fault power slider
The RQT window should now have a slider from 0.0 to 30.0 watts for the high power draw fault, allowing the user to choose a specific value for the fault injection. This value can be dynamically updated even while the fault is already injected, and the injection will apply the new value.

There is some leftover code of which I'm not sure yet if I should remove, comment out, or simply leave as-is despite its lack of use.