re-implemented previous fault injection
The old logic for injecting faults via inputs from a .CSV file has been uncommented again and re-added in a new statement to separate it from the high_power_draw fault (which works differently). Currently it shouldn't be called at all, but it is there if a future user wants to implement a new fault using the .CSV setup.