Implemented custom fault support
There is now a custom fault entry under dynamic reconfigure -> faults. A user can input the file path to their custom CSV file and activate it. Doing so locks the user out of using the provided high power draw fault for the rest of the simulation (and vice versa; using the high power draw locks the user out of using a custom fault). Additional warnings and informational outputs have been created to explain relevant behavior to the user.

There should be some error catching should the user attempt to designate an invalid file or directory, though nothing is stopping the user from designating a legitimate CSV file that is formatted improperly.