reworked custom fault activation
Instead of needing to enter the entire file path for a CSV fault file, the user can simply enter the file name and place it in the new '/ow_power_system/profiles' directory. Warnings and output were slightly adjusted to make this known to the user.

There is still an error involving re-activating the same custom fault, which will be addressed next.