Additional small fixes
As requested in previous comments:

The fault name and presentation in Faults.cfg has been changed.

Custom fault profiles now restart at index 0 if the simulation injects the fault past the end of the CSV.

The code for preventing combining high power draw with custom faults has been temporarily commented out, in anticipation of a discussion on whether or not it is needed.

The example_fault.csv length has been considerably shortened.