Addressed additional outputs & off-by-one errors
The warnings that high power draw cannot be combined with custom faults has now been properly commented out.

If the end of a custom fault profile is reached, the behavior now disables further injection of the fault until it is disabled and re-enabled (doing so restarts the index at the beginning). Outputs and user warnings updated to facilitate this.

There was previously an off-by-one error regarding custom fault injection. The simulation would skip over index 0. This should be fixed now.