First major milestone in multi-cell support
Pushing this to GitHub now that a breakthrough has seemingly been attained. This code is far from working, however, and there are many debug lines and even irrelevant files/functions. This is mostly as a backup for the current stage of work.

Using 8 MessageBuses, I've seemingly gotten the simulation to output 8 different EoD predictions in a single GSAP cycle. Next up is organizing the data and publishing only one set of information, which will require significant additional work. However, the foundation should be set.