Mechanical power values are now published
The raw and average mechanical values are now published at the same rate they were before the multi-cell model was implemented. According to Chetan, the published raw values should be averaged out, while all average values should be identical so any one of them will do.

However, all mechanical power values appear to be 0 at all times. This behavior is also seen in the noetic-devel branch, suggesting this is either expected behavior or a long-time bug.

There are many debug and test statements present that might stay through the merge to OW-994, but should be removed before the final merge to noetic-devel.