rewrote pack loop to address critical behavior
The old loop only sent input data to GSAP's prognosers whenever a new prediction needed to be triggered. This appears to have caused the invalid behavior seen in the previous commit, as the prognosers did not have enough input data to generate a valid prediction. This has been addressed, and the power system seems capable of handling 24 nodes at once now.

- The loop now sends input data each cycle to all prognosers still generating a prediction. Prognosers with completed predictions temporarily stop receiving data to prevent a new prediction from being triggered until all prognosers are done.

- To address the issue of having no predictions for the first several cycles when the initial batch of data is being processed, a set of initial data will be published at first:
-- RUL will be MAX_HORIZON_SECS
-- SoC will be 0.95 (95%)
-- TMP will be initial_temperature

-- Primary changes are in PowerSystemPack.cpp, but there are minor changes to other classes (such as modifying var types and removing now-unused functions to fit with the new changes).