Added OW-1173 changes
Moving away from system_clock calls turned out to be generally pointless after looking further into them (with the exception of one change outlined below), so OW-1173 is very minor and not worth a separate PR or review. The commit is being added to the OW-1163 PR because the changes are different for the multi-cell model than the single-cell model currently in noetic-devel.

- Updated the PrognoserInputHandler class's generateVoltageEstimate function to use a timelapse of ROS simulated time rather than system clock time, preventing potential desync.